Integrated decentralization adds to main (#60)
* Master: Integrated Simple Dependency Graph and Interrupts

* Client/Master: Added Support for Initial State Sending for Interrupts

* Master: Fixed Callback Support for interrupts.

* Added Divider

* Master: Repaired Performance Issues

* Client + Master: Implemented bitmap based interrupt configuration

* Client: Thread Shutdown on Disconnect

* client: Switch to jthread implementation for disconnect handling

* client: Switch abstract device cv to condition_variable_any and remaining threads to jthread

* Master + Protocol: Added basic exception handling for fatal errors.

* Client: Fixed Typo With Last Commit

* Master: Repaired Trigger Config for Interrupts

* lang: Add basic symbolic dependency graph

* Lang: Addded Divider Visitor

* Lang: Added Node Annotation of symgraph

* Symgraph + Divider: Updated Divider with the Code Division Algorithm

* Master: Added Scheduler

* Master: Added Scheduler

* Scheduler + Own Protocol : Progressed on Master and Client Implementation of own protocol.

* Scheduler: Implemented Scheduler

* Depgraph: Added production of oblock and metadata.

* Client: Added Gateway and Router

* Scheduler: Fixed out of order ownership calls

* common: Fix rebase errors

* libCE: Change DeviceHandleContainer to ManagedDevice

* drivers: Remove KEYBOARD driver

* Master + Client: Fixed integration bugs and added CONFIRM_OK to ownership protocol

* Client: Added Client Gateway for Decentralization

* Client: Added Client Gateway and Additional Packages

* Integrated Decentralization: Fixed Oblock ID System

* Master: Fixed Scheduler for multiple devices

* Master: Implemented Priority-Based scheduling for Interrupts triggered oblocks

* Scheduler: Implemented Triggering/Scheduleing  on callbacks and vtypes

* Master: Fixed building issues with old attributes

---------

+++ src/lang/libinterpreter/interpreter.cpp
@@ -360,24 +360,24 @@ BlsObject Interpreter::visit(AstNode::Expression::Unary& ast) {
 
         case UNARY_OPERATOR::INC:
             if (position == AstNode::Expression::Unary::OPERATOR_POSITION::PREFIX) {
-                object = object + 1;
+                object = object + int64_t(1);
                 return object;
             }
             else {
                 auto objCopy = object;
-                object = object + 1;
+                object = object + int64_t(1);
                 return objCopy;
             }
         break;
 
         case UNARY_OPERATOR::DEC:
             if (position == AstNode::Expression::Unary::OPERATOR_POSITION::PREFIX) {
-                object = object - 1;
+                object = object - int64_t(1);
                 return object;
             }
             else {
                 auto objCopy = object;
-                object = object - 1;
+                object = object - int64_t(1);
                 return objCopy;
             }
         break;
